+++ ams-desk-cs-backend/Deliveries/Dtos/DeliveryDto.cs
@@ -35,7 +35,7 @@ public record DeliveryDto
     public DateTime? StartDate { get; init; }
     public DateTime? FinishDate { get; init; }
     public int InvoiceId { get; init; }
-    public int PlaceId { get; init; }
+    public short PlaceId { get; init; }
     public DeliveryStatus Status { get; init; }
     public Place? Place { get; init; }
     public InvoiceDto? Invoice { get; init; }
