+++ ams-desk-cs-backend/Deliveries/Dtos/NewDeliveryDto.cs
@@ -4,7 +4,7 @@ namespace ams_desk_cs_backend.Deliveries.Dtos;
 
 public record NewDeliveryDto
 {
-    public int PlaceId { get; init; }
+    public short PlaceId { get; init; }
     
     public int InvoiceId { get; init; }
